pyproject.toml
[tool.poetry]
name = "netfone"
version = "0.1.0"
description = ""
authors = ["oozman <oomarusman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
black = "^24.8.0"
isort = "^5.13.2"
requests = "^2.32.3"
pygithub = "^2.4.0"
pytz = "^2024.2"
streamlit = "^1.38.0"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
